Collapse unreleased Execution API versions under 2026-06-30
The 2026-06-16 and 2026-06-30 Execution API versions were both added
after the 3.2.0 release (whose HEAD was 2026-04-06) and have not shipped
in any released Airflow version. Per the release process, unreleased
Cadwyn migrations between the last released version and the tentative
release date are collapsed before the release.

Merge the four VersionChange classes from 2026-06-16 (retry policy
fields, team_name field, assets-by-alias endpoint, task/asset state
store endpoints) into the single 2026-06-30 version, drop the
2026-06-16 version file, and point the version-gated tests at the new
previous version (2026-04-06). HEAD is unchanged, so the task SDK
generated datamodels (API_VERSION 2026-06-30) are unaffected.
