+++ airflow-core/src/airflow/api_fastapi/execution_api/versions/__init__.py
@@ -40,16 +40,14 @@ from airflow.api_fastapi.execution_api.versions.v2026_04_06 import (
     MovePreviousRunEndpoint,
     RemoveUpstreamMapIndexesField,
 )
-from airflow.api_fastapi.execution_api.versions.v2026_06_16 import (
-    AddAssetsByAliasEndpoint,
-    AddRetryPolicyFields,
-    AddTaskAndAssetStateStoreEndpoints,
-    AddTeamNameField,
-)
 from airflow.api_fastapi.execution_api.versions.v2026_06_30 import (
+    AddAssetsByAliasEndpoint,
     AddAwaitingInputStatePayload,
     AddConnectionTestEndpoint,
+    AddRetryPolicyFields,
+    AddTaskAndAssetStateStoreEndpoints,
     AddTaskInstanceQueueField,
+    AddTeamNameField,
     AddVariableKeysEndpoint,
 )
 
@@ -61,9 +59,6 @@ bundle = VersionBundle(
         AddConnectionTestEndpoint,
         AddAwaitingInputStatePayload,
         AddTaskInstanceQueueField,
-    ),
-    Version(
-        "2026-06-16",
         AddRetryPolicyFields,
         AddTeamNameField,
         AddTaskAndAssetStateStoreEndpoints,
