+++ airflow-core/tests/unit/api_fastapi/execution_api/versions/v2026_06_30/test_connection_tests.py
@@ -28,7 +28,7 @@ pytestmark = pytest.mark.db_test
 @pytest.fixture
 def old_ver_client(client):
     """Client configured to use API version before connection-tests endpoint was added."""
-    client.headers["Airflow-API-Version"] = "2026-06-16"
+    client.headers["Airflow-API-Version"] = "2026-04-06"
     return client
